Autogenerate Pokemon species and moves.
To do this, I also ended up attempting to convert the build system to Bazel.
Only time will tell if this was a good decision, as I'm not sure that the
Windows support is particularly mature; I may end up needing to go back to
a largely managed-by-hand method for generating the files.

Additionally, added a submodule pointing to a big database of Pokemon data,
which we can continue pulling from in the future to get even richer info.